+++ test/http/t54_tlsloop.cc
@@ -12,7 +12,7 @@
 #include <sstream>
 
 #define private public
-#include "galay-http2/kernel/stream_mgr.h"
+#include "galay-http2/kernel/stream_manager.h"
 #undef private
 
 using namespace galay::http2;