+++ test/http2/t51_h2cfast.cc
@@ -6,7 +6,7 @@
 #include <sstream>
 
 #define private public
-#include "galay-http2/kernel/stream_mgr.h"
+#include "galay-http2/kernel/stream_manager.h"
 #undef private
 #include "galay-http2/protoc/http2_frame.h"