+++ test/http2/t58_h2ack.cc
@@ -1,7 +1,7 @@
 #include <sstream>
 
 #define private public
-#include "galay-http2/kernel/stream_mgr.h"
+#include "galay-http2/kernel/stream_manager.h"
 #undef private
 
 #include <cassert>